Abstract FormField
