+++ src/components/Form/FormGroup.jsx
@@ -1,10 +1,13 @@
 import React from 'react';
 import styles from './FormGroup.css';
 
-export default ({ children, formName, className }) => (
+export default ({ children, formName, className, allSelectedElements, addChip, deleteChip }) => (
   <div className={`${styles.formGroup} ${className}`}>
     {React.Children.map(children, child => React.cloneElement(child, {
       formName,
+      selectedElements: [].concat(allSelectedElements[child.props.name] || []),
+      addChip,
+      deleteChip,
     }))}
   </div>
 );
